Vestuario/requirements.txt
pydantic
fastapi
pymongo
uvicorn
dotenv
python-jose
pytest
passlib
httpx
